# Back/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TypeContrat": {
				"enum": [
					"CDI",
					"CDD",
					"stage"
				],
				"type": "string"
			},
			"Role": {
				"enum": [
					"Admin",
					"RH",
					"Superviseur",
					"Employe"
				],
				"type": "string"
			},
			"UserOutput": {
				"properties": {
					"id_user": {
						"type": "number",
						"format": "double"
					},
					"matricule": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"empreinte": {
						"type": "string"
					},
					"poste": {
						"type": "string"
					},
					"type_contrat": {
						"$ref": "#/components/schemas/TypeContrat"
					},
					"date_embauche": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"date_fin_contrat": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"id_lieu": {
						"type": "number",
						"format": "double"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"id_departement": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"id_user",
					"matricule",
					"nom",
					"email",
					"phone",
					"badge",
					"poste",
					"type_contrat",
					"date_embauche",
					"id_lieu",
					"id_equipe",
					"id_departement",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserValidationResult": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.matricule-or-nom-or-prenom-or-email-or-phone-or-badge-or-empreinte-or-poste-or-type_contrat-or-date_embauche-or-date_fin_contrat-or-id_lieu-or-id_equipe-or-id_departement-or-role-or-password_": {
				"properties": {
					"matricule": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"empreinte": {
						"type": "string"
					},
					"poste": {
						"type": "string"
					},
					"type_contrat": {
						"$ref": "#/components/schemas/TypeContrat"
					},
					"date_embauche": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"date_fin_contrat": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"id_lieu": {
						"type": "number",
						"format": "double"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"id_departement": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"matricule",
					"nom",
					"email",
					"phone",
					"badge",
					"poste",
					"type_contrat",
					"date_embauche",
					"id_lieu",
					"id_equipe",
					"id_departement",
					"role"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserCreationParams": {
				"$ref": "#/components/schemas/Pick_IUser.matricule-or-nom-or-prenom-or-email-or-phone-or-badge-or-empreinte-or-poste-or-type_contrat-or-date_embauche-or-date_fin_contrat-or-id_lieu-or-id_equipe-or-id_departement-or-role-or-password_"
			},
			"Partial_UserCreationParams_": {
				"properties": {
					"matricule": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"empreinte": {
						"type": "string"
					},
					"poste": {
						"type": "string"
					},
					"type_contrat": {
						"$ref": "#/components/schemas/TypeContrat"
					},
					"date_embauche": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"date_fin_contrat": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						]
					},
					"id_lieu": {
						"type": "number",
						"format": "double"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"id_departement": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"password": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UserUpdateParams": {
				"$ref": "#/components/schemas/Partial_UserCreationParams_"
			},
			"IUser": {
				"properties": {
					"id_user": {
						"type": "number",
						"format": "double"
					},
					"matricule": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"prenom": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"empreinte": {
						"type": "string"
					},
					"poste": {
						"type": "string"
					},
					"type_contrat": {
						"$ref": "#/components/schemas/TypeContrat"
					},
					"date_embauche": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"date_fin_contrat": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"id_lieu": {
						"type": "number",
						"format": "double"
					},
					"id_equipe": {
						"type": "number",
						"format": "double"
					},
					"id_departement": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"id_user",
					"matricule",
					"nom",
					"email",
					"phone",
					"badge",
					"poste",
					"type_contrat",
					"date_embauche",
					"id_lieu",
					"id_equipe",
					"id_departement",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back-rh-connect",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutput"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateParams"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateUserBase",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserValidationResult"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/matricule/{matricule}": {
			"get": {
				"operationId": "GetUserByMatricule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserOutput"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "matricule",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/departement/{departementId}": {
			"get": {
				"operationId": "GetUsersByDepartement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "departementId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/equipe/{equipeId}": {
			"get": {
				"operationId": "GetUsersByEquipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/lieu/{id_lieu}": {
			"get": {
				"operationId": "GetUsersByLieu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH",
							"superviseur"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id_lieu",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/role/{role}": {
			"get": {
				"operationId": "GetUsersByRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserOutput"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "role",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Role"
						}
					}
				]
			}
		},
		"/users/import-api": {
			"post": {
				"operationId": "ImportFromApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"errors": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"totalUsers": {
											"type": "number",
											"format": "double"
										},
										"count": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"count",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"RH"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}